+++ autokuma/src/error.rs
@@ -14,6 +14,10 @@ pub enum Error {
     #[error(transparent)]
     Database(#[from] sled::Error),
 
+    #[cfg(feature = "kubernetes")]
+    #[error(transparent)]
+    K8S(#[from] K8SError),
+
     #[error("Error while trying to parse labels: {0}")]
     LabelParseError(String),
 
@@ -28,6 +32,13 @@ pub enum Error {
 
     #[error("No {} named {} could be found", .0.type_name(), .0.name())]
     NameNotFound(Name),
+}
+
+#[cfg(feature = "kubernetes")]
+#[derive(Error, Debug)]
+pub enum K8SError {
+    #[error("Finalizer Error: {0}")]
+    FinalizerError(#[source] Box<kube::runtime::finalizer::Error<Error>>),
 }
 
 pub type Result<T> = std::result::Result<T, Error>;