+++ autokuma/src/sources/docker_source.rs
@@ -15,9 +15,12 @@ use bollard::{
 };
 use itertools::Itertools;
 use kuma_client::util::ResultLogger;
-use std::{collections::HashMap, env};
+use std::{collections::HashMap, env, sync::Arc};
 
-async fn get_kuma_containers(state: &AppState, docker: &Docker) -> Result<Vec<ContainerSummary>> {
+async fn get_kuma_containers(
+    state: Arc<AppState>,
+    docker: &Docker,
+) -> Result<Vec<ContainerSummary>> {
     Ok(docker
         .list_containers(Some(ListContainersOptions::<String> {
             all: true,
@@ -44,7 +47,7 @@ async fn get_kuma_containers(state: &AppState, docker: &Docker) -> Result<Vec<Co
         .collect::<Vec<_>>())
 }
 
-async fn get_kuma_services(state: &AppState, docker: &Docker) -> Result<Vec<Service>> {
+async fn get_kuma_services(state: Arc<AppState>, docker: &Docker) -> Result<Vec<Service>> {
     Ok(docker
         .list_services(Some(ListServicesOptions::<String> {
             ..Default::default()
@@ -76,7 +79,7 @@ async fn get_kuma_services(state: &AppState, docker: &Docker) -> Result<Vec<Serv
 }
 
 fn get_entities_from_containers(
-    state: &AppState,
+    state: Arc<AppState>,
     system_info: &SystemInfo,
     containers: &Vec<ContainerSummary>,
 ) -> Result<HashMap<String, Entity>> {
@@ -98,16 +101,16 @@ fn get_entities_from_containers(
             template_values.insert("container", &container);
             template_values.insert("system_info", system_info);
 
-            let kuma_labels = get_kuma_labels(state, container.labels.as_ref(), &template_values)?;
+            let kuma_labels = get_kuma_labels(&state, container.labels.as_ref(), &template_values)?;
 
-            get_entities_from_labels(state, kuma_labels, &template_values)
+            get_entities_from_labels(state.clone(), kuma_labels, &template_values)
         })
         .flatten_ok()
         .try_collect()
 }
 
 fn get_entities_from_services(
-    state: &AppState,
+    state: Arc<AppState>,
     system_info: &SystemInfo,
     services: &Vec<Service>,
 ) -> Result<HashMap<String, Entity>> {
@@ -122,31 +125,47 @@ fn get_entities_from_services(
             let spec = service.spec.as_ref();
             let labels = spec.and_then(|spec| spec.labels.as_ref());
 
-            let kuma_labels = get_kuma_labels(state, labels, &template_values)?;
+            let kuma_labels = get_kuma_labels(&state, labels, &template_values)?;
 
-            get_entities_from_labels(state, kuma_labels, &template_values)
+            get_entities_from_labels(state.clone(), kuma_labels, &template_values)
         })
         .flatten_ok()
         .try_collect()
 }
 
-pub struct DockerSource {}
+pub struct DockerSource {
+    state: Arc<AppState>,
+}
 
 #[async_trait]
 impl Source for DockerSource {
-    async fn get_entities(&mut self, state: &AppState) -> Result<Vec<(String, Entity)>> {
-        if !state.config.docker.enabled {
+    fn name(&self) -> &'static str {
+        "Docker"
+    }
+
+    async fn init(&mut self) -> Result<()> {
+        Ok(())
+    }
+
+    async fn shutdown(&mut self) -> Result<()> {
+        Ok(())
+    }
+
+    async fn get_entities(&mut self) -> Result<Vec<(String, Entity)>> {
+        if !self.state.config.docker.enabled {
             return Ok(vec![]);
         }
 
-        let docker_hosts = state
+        let docker_hosts = self
+            .state
             .config
             .docker
             .hosts
             .clone()
             .map(|f| f.into_iter().map(Some).collect::<Vec<_>>())
             .unwrap_or_else(|| {
-                vec![state
+                vec![self
+                    .state
                     .config
                     .docker
                     .socket_path
@@ -170,25 +189,35 @@ impl Source for DockerSource {
 
             let system_info: bollard::secret::SystemInfo = docker.info().await.unwrap_or_default();
 
-            if state.config.docker.source == config::DockerSource::Containers
-                || state.config.docker.source == config::DockerSource::Both
+            if self.state.config.docker.source == config::DockerSource::Containers
+                || self.state.config.docker.source == config::DockerSource::Both
             {
-                let containers = get_kuma_containers(&state, &docker).await?;
+                let containers = get_kuma_containers(self.state.clone(), &docker).await?;
                 entities.extend(get_entities_from_containers(
-                    &state,
+                    self.state.clone(),
                     &system_info,
                     &containers,
                 )?);
             }
 
-            if state.config.docker.source == config::DockerSource::Services
-                || state.config.docker.source == config::DockerSource::Both
+            if self.state.config.docker.source == config::DockerSource::Services
+                || self.state.config.docker.source == config::DockerSource::Both
             {
-                let services = get_kuma_services(&state, &docker).await?;
-                entities.extend(get_entities_from_services(&state, &system_info, &services)?);
+                let services = get_kuma_services(self.state.clone(), &docker).await?;
+                entities.extend(get_entities_from_services(
+                    self.state.clone(),
+                    &system_info,
+                    &services,
+                )?);
             }
         }
 
         Ok(entities)
+    }
+}
+
+impl DockerSource {
+    pub fn new(state: Arc<AppState>) -> Self {
+        Self { state }
     }
 }